+++ cadhy/__init__.py
@@ -6,7 +6,7 @@ Toolkit for parametric modeling of hydraulic infrastructure within Blender.
 bl_info = {
     "name": "CADHY",
     "author": "CADHY Team",
-    "version": (0, 2, 1),
+    "version": (0, 2, 2),
     "blender": (4, 1, 0),
     "location": "View3D > Sidebar > CADHY",
     "description": "Parametric modeling toolkit for hydraulic infrastructure and CFD domain generation",